pharmaserver/requirements.txt
Django==1.9.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.10.0
djangorestframework==3.3.3
djoser==0.5.1
docutils==0.12
drf-extensions==0.2.8
oauthlib==1.0.3
Pillow==3.4.2
six==1.10.0
